requirements.txt
# SHEIN_SEN - Système d'Automatisation des Commandes Groupées Shein
# Dépendances Python

# WhatsApp API
twilio==8.10.0
requests==2.31.0

# IA et traitement de données
openai==1.3.7
re
json

# Automatisation navigateur
playwright==1.40.0
selenium==4.15.2

# Manipulation de fichiers
pandas==2.1.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# PDF
reportlab==4.0.7
fpdf2==2.7.6

# Google Sheets (optionnel)
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Interface web (optionnel)
flask==3.0.0
streamlit==1.28.2
fastapi==0.104.1
uvicorn==0.24.0

# Utilitaires
python-dotenv==1.0.0
loguru==0.7.2
beautifulsoup4==4.12.2
lxml==4.9.3
validators==0.22.0